separate initEditModeControls into a function for squareedit which initializes the panels on the right, and a general function which gets called in NeonView and which sets up the button and dropdown menu in the top left
