+++ src/SquareEdit/Controls.js
@@ -5,6 +5,21 @@ import { unselect } from '../utils/SelectTools.js';
 const $ = require('jquery');
 
 /**
+ * Set listener on EditMode button.
+ * @param {EditMode} editMode - The EditMode object.
+ */
+export function initEditModeControls (editMode) {
+  /* document.getElementById('dropdown_toggle').innerHTML =
+    '<a class="navbar-item"><button class="button" id="edit_mode">' +
+    'Edit MEI</button></a>'; */
+  $('#edit_mode').on('click', function () {
+    $('#insert_controls').append(Contents.insertControlsPanel);
+    $('#edit_controls').append(Contents.editControlsPanel);
+    editMode.initEditMode();
+  });
+}
+
+/**
  * Bind listeners to insert tabs.'
  * @param {InsertHandler} insertHandler - An InsertHandler to run the tasks.
  */
